@@ -3,7 +3,7 @@
     <PropertyGroup>
         <OutputType>WinExe</OutputType>
         <IsPackable>true</IsPackable>
-        <TargetFrameworks>netcoreapp3.1;net5.0-windows;</TargetFrameworks>
+        <TargetFrameworks>net6.0-windows;net5.0-windows;netcoreapp3.1;</TargetFrameworks>
         <UseWindowsForms>true</UseWindowsForms>
         <ApplicationIcon>icon.ico</ApplicationIcon>
         <Description>Liyanjie.AspNetCore.Hosting.WindowsDesktop</Description>
@@ -21,22 +21,28 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <None Remove="Host\net5.0-windows\Host.deps.json" />
-        <None Remove="Host\net5.0-windows\Host.dll" />
-        <None Remove="Host\net5.0-windows\Host.dll.config" />
-        <None Remove="Host\net5.0-windows\Host.exe" />
-        <None Remove="Host\net5.0-windows\Host.runtimeconfig.json" />
-        <None Remove="Host\net5.0-windows\ref\Host.dll" />
-        <None Remove="Host\net5.0-windows\wwwroot\index.html" />
-        <None Remove="Host\netcoreapp3.1\Host.deps.json" />
-        <None Remove="Host\netcoreapp3.1\Host.dll" />
-        <None Remove="Host\netcoreapp3.1\Host.dll.config" />
-        <None Remove="Host\netcoreapp3.1\Host.exe" />
-        <None Remove="Host\netcoreapp3.1\Host.runtimeconfig.json" />
-        <None Remove="Host\netcoreapp3.1\wwwroot\index.html" />
-    </ItemGroup>
-
-    <ItemGroup>
+		<Content Include="Host\net6.0-windows\Host.deps.json">
+			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
+		</Content>
+		<Content Include="Host\net6.0-windows\Host.dll">
+			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
+		</Content>
+		<Content Include="Host\net6.0-windows\Host.dll.config" />
+		<Content Include="Host\net6.0-windows\Host.exe">
+			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
+		</Content>
+		<Content Include="Host\net6.0-windows\Host.runtimeconfig.json">
+			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
+		</Content>
+		<Content Include="Host\net6.0-windows\Host.staticwebassets.runtime.json">
+			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
+		</Content>
+		<Content Include="Host\net6.0-windows\ref\Host.dll">
+			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
+		</Content>
+		<Content Include="Host\net6.0-windows\wwwroot\index.html">
+			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
+		</Content>
         <Content Include="Host\net5.0-windows\Host.deps.json">
             <CopyToOutputDirectory>Always</CopyToOutputDirectory>
         </Content>
