@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Configuration;
 using System.IO;
-using System.Linq;
 using System.Reflection;
 using System.Threading;
 using System.Windows.Forms;
@@ -14,29 +13,29 @@ namespace Liyanjie.AspNetCore.Hosting.WindowsDesktop
 {
     class HostManager
     {
-        static readonly string[] startup;
-        static readonly string[] urls;
-        static IHost host;
+        static readonly string _contentRoot;
+        static readonly string _webRoot;
+        static readonly Type _startupType;
+        static readonly string[] _urls;
+
         static CancellationTokenSource cts;
+        static IHost host;
 
         static HostManager()
         {
             try
             {
-                startup = ConfigurationManager.AppSettings["Startup"]?.Split(',', StringSplitOptions.RemoveEmptyEntries);
-                urls = ConfigurationManager.AppSettings["Urls"]?.Split(',', StringSplitOptions.RemoveEmptyEntries);
+                _contentRoot = ConfigurationManager.AppSettings["ContentRoot"];
+                _webRoot = ConfigurationManager.AppSettings["WebRoot"];
+                var startup = ConfigurationManager.AppSettings["Startup"]?.Split(',', StringSplitOptions.RemoveEmptyEntries);
+                _startupType = Assembly.LoadFrom(startup[0]).GetType(startup[1]);
+                _urls = ConfigurationManager.AppSettings["Urls"]?.Split(',', StringSplitOptions.RemoveEmptyEntries);
             }
             catch { }
         }
 
         internal static void Start()
         {
-            if (startup is null || startup.Length < 2)
-            {
-                MessageBox.Show("AppSettings\\Startup 配置错误");
-                return;
-            }
-
             foreach (var item in Directory.EnumerateFiles(AppDomain.CurrentDomain.BaseDirectory, "*.dll", new EnumerationOptions
             {
                 RecurseSubdirectories = true,
@@ -48,18 +47,31 @@ namespace Liyanjie.AspNetCore.Hosting.WindowsDesktop
                 }
                 catch { }
             }
+
+            void ConfigureWebHost(IWebHostBuilder webBuilder)
+            {
+                if (!string.IsNullOrWhiteSpace(_contentRoot))
+                    webBuilder.UseWebRoot(_contentRoot);
+                if (!string.IsNullOrWhiteSpace(_webRoot))
+                    webBuilder.UseWebRoot(_webRoot);
+                webBuilder.UseStartup(_startupType);
+                if (_urls?.Length > 0)
+                    webBuilder.UseUrls(_urls);
+                webBuilder.ConfigureLogging(logging => logging.AddProvider(new Logging.MyLoggerProvider()));
+            }
+
             try
             {
                 cts = new();
-                host = Host.CreateDefaultBuilder(Environment.GetCommandLineArgs())
-                    .ConfigureWebHostDefaults(webBuilder =>
-                    {
-                        webBuilder.UseStartup(Assembly.LoadFrom(startup[0]).GetType(startup[1]));
-                        if (urls?.Length > 0)
-                            webBuilder.UseUrls(urls);
-                        webBuilder.ConfigureLogging(logging => logging.AddProvider(new Logging.MyLoggerProvider()));
-                    })
+#if NET6_0_OR_GREATER
+                var builder = Microsoft.AspNetCore.Builder.WebApplication.CreateBuilder();
+                ConfigureWebHost(builder.WebHost);
+                host=builder.Build();
+#else
+                host = Host.CreateDefaultBuilder()
+                    .ConfigureWebHostDefaults(ConfigureWebHost)
                     .Build();
+#endif
                 host.RunAsync(cts.Token);
             }
             catch (Exception ex)
@@ -74,7 +86,7 @@ namespace Liyanjie.AspNetCore.Hosting.WindowsDesktop
             {
                 if (host is not null)
                 {
-                    host.StopAsync(cts.Token).ConfigureAwait(false);
+                    host.StopAsync().ConfigureAwait(false);
                     host.Dispose();
                 }
             }
@@ -85,7 +97,7 @@ namespace Liyanjie.AspNetCore.Hosting.WindowsDesktop
         }
         internal static string[] GetUrls()
         {
-            return urls?.Length > 0 ? urls : new[] { "http://localhost:5000" };
+            return _urls?.Length > 0 ? _urls : new[] { "http://localhost:5000" };
         }
     }
 }
